moved validation into seperate script so it can be ran independently of the save functionality
