+++ ingest.py
@@ -1,13 +1,13 @@
 import pandas as pd
 import numpy as np
 import sqlalchemy as db
-from jsonschema import validate, Draft3Validator
-from jsonschema.exceptions import ValidationError, SchemaError 
+ 
 import os
 import sys
-from json import load
 
 import logging
+
+from validate import verify
 logging.basicConfig()
 logging.getLogger("sqlalchemy.engine").setLevel(logging.ERROR)
 logging.getLogger("sqlalchemy.pool").setLevel(logging.ERROR)
@@ -19,15 +19,6 @@ db_prod_ip=os.environ.get('DB_PROD_IP')
 db_port=os.environ.get('DB_PORT')
 db_name=os.environ.get('DB_NAME')
 
-schemas = {
-    'chemdoseResponseVals': load(open('./schemas/chemdoseResponseVals.json')),
-    'chemicalsByExtractSample': load(open('./schemas/chemicalByExtractSample.json')),
-    'chemSummaryStats': load(open('./schemas/chemSummaryStats.json')),
-    'chemXYcoords': load(open('./schemas/chemXYcoords.json')),
-    'envSampdoseResponseVals': load(open('./schemas/envSampdoseResponseVals.json')),
-    'envSampSummaryStats': load(open('./schemas/envSampSummaryStats.json')),
-    'envSampXYcoords': load(open('./schemas/envSampXYcoords.json')),
-}
 
 def test_connection(database):
     if database == "production":
@@ -88,39 +79,10 @@ def read_and_save(csv_file, table_name, if_exists, engine):
     df.replace([np.inf, -np.inf], np.nan, inplace=True)
     valid = verify(df, table_name)
     if valid:
-        # TODO uncomment once done testing schema
-        #df.to_sql(table_name, engine, if_exists=if_exists, index=False)
+        df.to_sql(table_name, engine, if_exists=if_exists, index=False)
         print("\tFinished writing to {}.".format(table_name))
     else:
         print("Invalid schema, not saving.")
-    
-
-
-def verify(df, table_name):
-    """ Takes a Pandas DataFrame and a Microsoft SQL Server table name to compare the number of rows in each. This function only works if the to_sql function replaces the table (if it appends the number of rows will obviously be off)
-
-    Parameters: 
-        csv_df (Pandas Dataframe): dataframe of the original data from CSV
-        table_name (string): name of table that you want to compare against
-
-    Returns:
-        nothing
-    """
-    print("\tVerifying schema...")
-    df = df.where(pd.notnull(df), None)
-    v = Draft3Validator(schemas[table_name])
-    errors = set()
-    for row in df.to_dict(orient='records'):
-        for error in sorted(v.iter_errors(row), key=str):
-            errors.add(str(error))   
-
-    if errors:
-        print('Validation errors when running schema check on {}'.format(table_name))
-        with open("/tmp/{}_validation_errors.txt".format(table_name), 'w+') as fp:
-            for error in errors:
-                fp.write("{}\n\n\n".format(error))
-        return False
-    return True
 
 
 def pull(table_name, engine):
